refactor(java): refactor Java file icons to be optionally styled
