+++ src/main/kotlin/com/github/catppuccin/jetbrains_icons/IconProvider.kt
@@ -1,13 +1,8 @@
 package com.github.catppuccin.jetbrains_icons
 
-import com.github.catppuccin.jetbrains_icons.settings.PluginSettingsState
-import com.github.catppuccin.jetbrains_icons.util.PsiClassUtils
 import com.intellij.ide.IconProvider
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
-import com.intellij.psi.JavaRecursiveElementVisitor
 import com.intellij.psi.util.PsiUtilCore
 import javax.swing.Icon
 
@@ -16,9 +11,9 @@ class IconProvider : IconProvider() {
 
     private val fileTypes = mapOf(
         "Dockerfile" to icons.docker,
-    ) + provideJavaIcons()
+    )
 
-    override fun getIcon(element: PsiElement, flags: Int): Icon {
+    override fun getIcon(element: PsiElement, flags: Int): Icon? {
         val virtualFile = PsiUtilCore.getVirtualFile(element)
 
         // File types
@@ -26,8 +21,12 @@ class IconProvider : IconProvider() {
             val file = PsiManager.getInstance(element.project).findFile(virtualFile)
             if (file != null) {
                 val fileType = when {
-                    file.name.endsWith(".java") -> determineJavaFileType(file)
-                    else -> file.fileType.name
+                   file.name.endsWith(".java") -> {
+                       return null
+                   }
+                    else -> {
+                        file.fileType.name
+                    }
                 }
                 val icon = fileTypes[fileType]
                 if (icon != null) {
@@ -67,38 +66,4 @@ class IconProvider : IconProvider() {
         return icons._file
     }
 
-
-    private fun determineJavaFileType(file: PsiFile): String {
-        var fileType = "JAVA_CLASS"
-        if (!PluginSettingsState.instance.javaSupport) return fileType
-        file.accept(object : JavaRecursiveElementVisitor() {
-            override fun visitClass(aClass: PsiClass) {
-                when {
-                    aClass.isAnnotationType -> fileType = "JAVA_ANNOTATION"
-                    aClass.isInterface -> fileType = "JAVA_INTERFACE"
-                    aClass.isEnum -> fileType = "JAVA_ENUM"
-                    aClass.isRecord -> fileType = "JAVA_RECORD"
-                    PsiClassUtils.isException(aClass) -> fileType = "JAVA_EXCEPTION"
-                    PsiClassUtils.isSealed(aClass) -> fileType = "JAVA_SEALED"
-                    PsiClassUtils.isFinal(aClass) -> fileType = "JAVA_FINAL"
-                    PsiClassUtils.isAbstract(aClass) -> fileType = "JAVA_ABSTRACT"
-                }
-            }
-        })
-        return fileType
-    }
-
-    private fun provideJavaIcons(): Map<String, Icon> {
-        return mapOf(
-            "JAVA_ANNOTATION" to icons.java_annotation,
-            "JAVA_INTERFACE" to icons.java_interface,
-            "JAVA_ENUM" to icons.java_enum,
-            "JAVA_RECORD" to icons.java_record,
-            "JAVA_EXCEPTION" to icons.java_exception,
-            "JAVA_ABSTRACT" to icons.java_class_abstract,
-            "JAVA_SEALED" to icons.java_class_sealed,
-            "JAVA_FINAL" to icons.java_class_final,
-            "JAVA_CLASS" to icons.java_class,
-        )
-    }
 }
