+++ src/main/resources/META-INF/com.github.catppuccin.jetbrains_icons-withJava.xml
@@ -1,0 +1,9 @@
+<idea-plugin>
+  <extensions defaultExtensionNs="com.intellij">
+    <iconProvider
+      implementation="com.github.catppuccin.jetbrains_icons.providers.JavaIconProvider"
+      order="first"
+      id="JavaIconProvider"
+    />
+  </extensions>
+</idea-plugin>
