+++ src/main/resources/META-INF/com.github.catppuccin.jetbrains_icons-withKotlin.xml
@@ -1,0 +1,2 @@
+<idea-plugin>
+</idea-plugin>
