+++ src/main/resources/META-INF/plugin.xml
@@ -31,8 +31,8 @@ For further help, see also:
 <p>LICENSE: <a href="https://github.com/catppuccin/jetbrains-icons/blob/main/LICENSE">MIT</a></p>
 ]]></description>
   <depends>com.intellij.modules.platform</depends>
-  <depends optional="true">org.jetbrains.kotlin</depends>
-  <depends optional="true">com.intellij.java</depends>
+  <depends optional="true" config-file="com.github.catppuccin.jetbrains_icons-withKotlin.xml">org.jetbrains.kotlin</depends>
+  <depends optional="true" config-file="com.github.catppuccin.jetbrains_icons-withJava.xml">com.intellij.java</depends>
   <extensions defaultExtensionNs="com.intellij">
     <iconProvider
       implementation="com.github.catppuccin.jetbrains_icons.IconProvider"
